refactor: #105 잘못된 DTO 위치 변경
- 추후 패키지 구조 변경 예정
