+++ src/main/java/org/thisway/member/dto/response/CompanyChefMembersResponse.java
@@ -2,6 +2,7 @@ package org.thisway.member.dto.response;
 
 import java.util.List;
 import org.thisway.common.PageInfo;
+import org.thisway.member.dto.CompanyChefMembersOutput;
 
 public record CompanyChefMembersResponse(
         List<CompanyChefMemberDetailResponse> members,
