+++ src/test/java/org/thisway/member/service/CompanyChefMemberServiceTest.java
@@ -25,7 +25,7 @@ import org.thisway.company.support.CompanyFixture;
 import org.thisway.member.dto.CompanyChefMemberDetailOutput;
 import org.thisway.member.dto.CompanyChefMemberRegisterInput;
 import org.thisway.member.dto.CompanyChefMemberUpdateInput;
-import org.thisway.member.dto.response.CompanyChefMembersOutput;
+import org.thisway.member.dto.CompanyChefMembersOutput;
 import org.thisway.member.entity.Member;
 import org.thisway.member.entity.MemberRole;
 import org.thisway.member.repository.MemberRepository;
